+++ Image Feed/Auth/WebViewController.swift
@@ -1,0 +1,15 @@
+import UIKit
+import WebKit
+
+enum WebViewConstants {
+    static let unsplashAuthorizeURLString = "https://unsplash.com/oauth/authorize"
+}
+
+final class WebViewViewController: UIViewController {
+    @IBOutlet weak var webView: WKWebView!
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        webView.backgroundColor = .ypWhite_1
+        
+    }
+}
