--- Image Feed/Auth/WebViewViewController.swift
@@ -1,73 +1,0 @@
-import UIKit
-import WebKit
-
-enum WebViewConstants {
-    static let unsplashAuthorizeURLString = "https://unsplash.com/oauth/authorize"
-}
-
-protocol WebViewViewControllerDelegate: AnyObject {
-    func webViewViewController(_ vc: WebViewViewController, didAuthenticateWithCode code: String)
-    func webViewViewControllerDidCancel(_ vc: WebViewViewController)
-}
-
-final class WebViewViewController:UIViewController {
-    @IBOutlet weak var webView: WKWebView!
-    @IBOutlet weak var backButton: UIButton!
-    @IBOutlet weak var progressView: UIProgressView!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        webView.navigationDelegate = self
-        webView.backgroundColor = .ypWhite_1
-        loadAuthURL()
-    }
-    @IBAction func buttonDidTapped(_ sender: Any) {
-        dismiss(animated: true, completion: nil)
-    }
-}
-
-extension WebViewViewController {
-    private func loadAuthURL() {
-        guard var urlComponents = URLComponents(string: WebViewConstants.unsplashAuthorizeURLString) else { return }
-        
-        urlComponents.queryItems = [
-            URLQueryItem(name: "client_id", value: Constants.accessKey),
-            URLQueryItem(name: "redirect_uri", value: Constants.redirectURI),
-            URLQueryItem(name: "response_type", value: "code"),
-            URLQueryItem(name: "scope", value: Constants.accessScope)
-        ]
-        
-        guard let url = urlComponents.url else { return }
-        let request = URLRequest(url: url)
-        webView.load(request)
-    }
-}
-
-extension WebViewViewController: WKNavigationDelegate {
-    func webView(
-        _ webView: WKWebView,
-        decidePolicyFor navigationAction: WKNavigationAction,
-        decisionHandler: @escaping (WKNavigationActionPolicy) -> Void
-    ) {
-         if let code = code(from: navigationAction) {
-                //TODO: process code                     //2
-                decisionHandler(.cancel) //3
-          } else {
-                decisionHandler(.allow) //4
-            }
-    }
-    
-    private func code(from navigationAction: WKNavigationAction) -> String? {
-        if
-            let url = navigationAction.request.url,
-            let urlComponents = URLComponents(string: url.absoluteString),
-            urlComponents.path == "/oauth/authorize/native",
-            let items = urlComponents.queryItems,
-            let codeItem = items.first(where: { $0.name == "code" })
-        {
-            return codeItem.value
-        } else {
-            return nil
-        }
-    }
-}
